introducing configure.Debug
Former-commit-id: 5d18cce5374a57d5f875d88519f4c39fa60ab265

+++ user/main/dd.c
@@ -114,7 +114,7 @@ int main(int argc, char *argv[])
                         sf_shortwrite(sbuf,nout,out);
 			break;
 		    case SF_INT:
-                        ibuf = (short*) bufout;
+                        ibuf = (int*) bufout;
                         for (i=j=0; i < nin && j < nout; i++, j++) {
 			    ibuf[j] = sbuf[i]; 
 			}
